// server/src/auth/enums/auth-error-code.enum.ts
export enum AuthErrorCode {
  GOOGLE_AUTH_REQUIRED = 'GOOGLE_AUTH_REQUIRED',
  GOOGLE_ACCOUNT_EXISTS = 'GOOGLE_ACCOUNT_EXISTS',
  ORGANIZATION_SUSPENDED = 'ORGANIZATION_SUSPENDED',
  ORGANIZATION_REQUIRED = 'ORGANIZATION_REQUIRED',
  USER_SUSPENDED = 'USER_SUSPENDED',
  REGISTRATION_DISABLED = 'REGISTRATION_DISABLED',
  OAUTH_CODE_INVALID = 'OAUTH_CODE_INVALID',
}